Revamp and use new `optional` attribute for `additional_rules`
The `optional` attribute is officially available starting from
terraform 1.3.0, which's not yet released.

If `terraform 1.3.0` is still not realeased at the moment you're using
this module, you could use one of the `beta/rc` releases of
`terraform 1.3.0` (`tfenv` is your friend).

Since I'm using `default` with `optional`, this will not work with
`terraform 1.2.8`, which's the latest stable version ATM, even if you
enable the experimental `module_variable_optional_attrs`. Maybe you
can wait for a future `1.2.9`, if there's any, but I'm not sure if
they are planning to include it in the experiment.

I pushed this experimental feature based on `claranet` team's request.